Put test classes into clear namespaces, so that they don't pollute the API.
