+++ core/src/main/kotlin/net/corda/core/schemas/testing/DummyDealStateSchemaV1.kt
@@ -1,4 +1,4 @@
-package net.corda.core.schemas
+package net.corda.core.schemas.testing
 
 /**
  * An object used to fully qualify the [DummyDealStateSchema] family name (i.e. independent of version).
@@ -9,7 +9,7 @@ object DummyDealStateSchema
  * First version of a cash contract ORM schema that maps all fields of the [DummyDealState] contract state as it stood
  * at the time of writing.
  */
-object DummyDealStateSchemaV1 : net.corda.core.schemas.MappedSchema(schemaFamily = net.corda.core.schemas.DummyDealStateSchema.javaClass, version = 1, mappedTypes = listOf(net.corda.core.schemas.DummyDealStateSchemaV1.PersistentDummyDealState::class.java)) {
+object DummyDealStateSchemaV1 : net.corda.core.schemas.MappedSchema(schemaFamily = net.corda.core.schemas.testing.DummyDealStateSchema.javaClass, version = 1, mappedTypes = listOf(net.corda.core.schemas.testing.DummyDealStateSchemaV1.PersistentDummyDealState::class.java)) {
     @javax.persistence.Entity
     @javax.persistence.Table(name = "dummy_deal_states")
     class PersistentDummyDealState(
