+++ core/src/main/kotlin/net/corda/core/schemas/testing/DummyLinearStateSchemaV1.kt
@@ -1,5 +1,7 @@
-package net.corda.core.schemas
+package net.corda.core.schemas.testing
 
+import net.corda.core.schemas.MappedSchema
+import net.corda.core.schemas.PersistentState
 import java.time.Instant
 import java.util.*
 import javax.persistence.Column
