+++ core/src/test/kotlin/net/corda/core/contracts/TransactionGraphSearchTests.kt
@@ -1,5 +1,7 @@
 package net.corda.core.contracts
 
+import net.corda.core.contracts.testing.DummyContract
+import net.corda.core.contracts.testing.DummyState
 import net.corda.core.crypto.newSecureRandom
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.WireTransaction
@@ -10,7 +12,6 @@ import net.corda.testing.MEGA_CORP_PUBKEY
 import net.corda.testing.node.MockServices
 import net.corda.testing.node.MockTransactionStorage
 import org.junit.Test
-import java.security.KeyPair
 import kotlin.test.assertEquals
 
 class TransactionGraphSearchTests {
