+++ core/src/test/kotlin/net/corda/core/contracts/TransactionTests.kt
@@ -1,6 +1,7 @@
 package net.corda.core.contracts
 
 import net.corda.contracts.asset.DUMMY_CASH_ISSUER_KEY
+import net.corda.core.contracts.testing.DummyContract
 import net.corda.core.crypto.CompositeKey
 import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.generateKeyPair
