+++ core/src/test/kotlin/net/corda/core/contracts/clauses/VerifyClausesTests.kt
@@ -1,6 +1,10 @@
 package net.corda.core.contracts.clauses
 
-import net.corda.core.contracts.*
+import net.corda.core.contracts.AuthenticatedObject
+import net.corda.core.contracts.CommandData
+import net.corda.core.contracts.ContractState
+import net.corda.core.contracts.TransactionForContract
+import net.corda.core.contracts.testing.DummyContract
 import net.corda.core.crypto.SecureHash
 import org.junit.Test
 import kotlin.test.assertFailsWith
