+++ core/src/test/kotlin/net/corda/core/flows/ContractUpgradeFlowTest.kt
@@ -3,6 +3,8 @@ package net.corda.core.flows
 import co.paralleluniverse.fibers.Suspendable
 import net.corda.contracts.asset.Cash
 import net.corda.core.contracts.*
+import net.corda.core.contracts.testing.DummyContract
+import net.corda.core.contracts.testing.DummyContractV2
 import net.corda.core.crypto.SecureHash
 import net.corda.core.getOrThrow
 import net.corda.core.identity.AbstractParty
