+++ core/src/test/kotlin/net/corda/core/flows/ResolveTransactionsFlowTest.kt
@@ -1,7 +1,6 @@
 package net.corda.core.flows
 
-import net.corda.core.contracts.DummyContract
-import net.corda.core.crypto.NullSignature
+import net.corda.core.contracts.testing.DummyContract
 import net.corda.core.crypto.SecureHash
 import net.corda.core.getOrThrow
 import net.corda.core.identity.Party
