+++ finance/src/main/java/net/corda/contracts/JavaCommercialPaper.java
@@ -1,27 +1,36 @@
 package net.corda.contracts;
 
-import co.paralleluniverse.fibers.*;
-import com.google.common.collect.*;
-import kotlin.*;
-import net.corda.contracts.asset.*;
+import co.paralleluniverse.fibers.Suspendable;
+import com.google.common.collect.ImmutableList;
+import kotlin.Pair;
+import kotlin.Unit;
+import net.corda.contracts.asset.CashKt;
 import net.corda.core.contracts.*;
-import net.corda.core.contracts.Contract;
-import net.corda.core.contracts.TransactionForContract.*;
-import net.corda.core.contracts.clauses.*;
-import net.corda.core.crypto.*;
+import net.corda.core.contracts.TransactionForContract.InOutGroup;
+import net.corda.core.contracts.clauses.AnyOf;
+import net.corda.core.contracts.clauses.Clause;
+import net.corda.core.contracts.clauses.ClauseVerifier;
+import net.corda.core.contracts.clauses.GroupClauseVerifier;
+import net.corda.core.crypto.SecureHash;
+import net.corda.core.crypto.testing.NullPublicKey;
 import net.corda.core.identity.AbstractParty;
 import net.corda.core.identity.AnonymousParty;
 import net.corda.core.identity.Party;
-import net.corda.core.node.services.*;
-import net.corda.core.transactions.*;
-import org.jetbrains.annotations.*;
+import net.corda.core.node.services.VaultService;
+import net.corda.core.transactions.TransactionBuilder;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-import java.time.*;
-import java.util.*;
-import java.util.stream.*;
+import java.time.Instant;
+import java.util.Collections;
+import java.util.Currency;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
-import static kotlin.collections.CollectionsKt.*;
-import static net.corda.core.contracts.ContractsDSL.*;
+import static kotlin.collections.CollectionsKt.single;
+import static net.corda.core.contracts.ContractsDSL.requireSingleCommand;
+import static net.corda.core.contracts.ContractsDSL.requireThat;
 
 
 /**
