+++ finance/src/main/kotlin/net/corda/contracts/DummyDealContract.kt
@@ -3,13 +3,14 @@ package net.corda.contracts
 import net.corda.core.contracts.Contract
 import net.corda.core.contracts.TransactionForContract
 import net.corda.core.contracts.UniqueIdentifier
-import net.corda.core.crypto.*
+import net.corda.core.crypto.SecureHash
+import net.corda.core.crypto.containsAny
 import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.Party
-import net.corda.core.schemas.DummyDealStateSchemaV1
 import net.corda.core.schemas.MappedSchema
 import net.corda.core.schemas.PersistentState
 import net.corda.core.schemas.QueryableState
+import net.corda.core.schemas.testing.DummyDealStateSchemaV1
 import net.corda.core.transactions.TransactionBuilder
 import java.security.PublicKey
