+++ finance/src/main/kotlin/net/corda/contracts/testing/VaultFiller.kt
@@ -7,6 +7,7 @@ import net.corda.contracts.DealState
 import net.corda.contracts.DummyDealContract
 import net.corda.contracts.asset.*
 import net.corda.core.contracts.*
+import net.corda.core.contracts.testing.DummyLinearContract
 import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.AnonymousParty
 import net.corda.core.identity.Party
