+++ finance/src/test/kotlin/net/corda/contracts/testing/Generators.kt
@@ -9,7 +9,7 @@ import net.corda.core.contracts.Command
 import net.corda.core.contracts.CommandData
 import net.corda.core.contracts.ContractState
 import net.corda.core.contracts.TransactionType
-import net.corda.core.crypto.NullSignature
+import net.corda.core.crypto.testing.NullSignature
 import net.corda.core.identity.AnonymousParty
 import net.corda.core.testing.*
 import net.corda.core.transactions.SignedTransaction
