+++ node/src/integration-test/kotlin/net/corda/node/services/RaftNotaryServiceTests.kt
@@ -1,10 +1,10 @@
 package net.corda.node.services
 
 import com.google.common.util.concurrent.Futures
-import net.corda.core.contracts.DummyContract
 import net.corda.core.contracts.StateAndRef
 import net.corda.core.contracts.StateRef
 import net.corda.core.contracts.TransactionType
+import net.corda.core.contracts.testing.DummyContract
 import net.corda.core.getOrThrow
 import net.corda.core.identity.Party
 import net.corda.core.map
