+++ node/src/test/kotlin/net/corda/node/services/NotaryChangeTests.kt
@@ -1,6 +1,7 @@
 package net.corda.node.services
 
 import net.corda.core.contracts.*
+import net.corda.core.contracts.testing.DummyContract
 import net.corda.core.crypto.generateKeyPair
 import net.corda.core.getOrThrow
 import net.corda.core.identity.Party
