+++ node/src/test/kotlin/net/corda/node/services/database/HibernateConfigurationTest.kt
@@ -10,9 +10,9 @@ import net.corda.core.contracts.*
 import net.corda.core.crypto.toBase58String
 import net.corda.core.node.services.Vault
 import net.corda.core.node.services.VaultService
-import net.corda.core.schemas.DummyLinearStateSchemaV1
-import net.corda.core.schemas.DummyLinearStateSchemaV2
 import net.corda.core.schemas.PersistentStateRef
+import net.corda.core.schemas.testing.DummyLinearStateSchemaV1
+import net.corda.core.schemas.testing.DummyLinearStateSchemaV2
 import net.corda.core.serialization.deserialize
 import net.corda.core.serialization.storageKryo
 import net.corda.core.transactions.SignedTransaction
