+++ node/src/test/kotlin/net/corda/node/services/database/RequeryConfigurationTest.kt
@@ -3,12 +3,12 @@ package net.corda.node.services.database
 import io.requery.Persistable
 import io.requery.kotlin.eq
 import io.requery.sql.KotlinEntityDataStore
-import net.corda.core.contracts.DummyContract
 import net.corda.core.contracts.StateRef
 import net.corda.core.contracts.TransactionType
+import net.corda.core.contracts.testing.DummyContract
 import net.corda.core.crypto.DigitalSignature
-import net.corda.core.crypto.NullPublicKey
 import net.corda.core.crypto.SecureHash
+import net.corda.core.crypto.testing.NullPublicKey
 import net.corda.core.crypto.toBase58String
 import net.corda.core.identity.AnonymousParty
 import net.corda.core.node.services.Vault
