+++ node/src/test/kotlin/net/corda/node/services/persistence/DBTransactionStorageTests.kt
@@ -3,8 +3,8 @@ package net.corda.node.services.persistence
 import net.corda.core.contracts.StateRef
 import net.corda.core.contracts.TransactionType
 import net.corda.core.crypto.DigitalSignature
-import net.corda.core.crypto.NullPublicKey
 import net.corda.core.crypto.SecureHash
+import net.corda.core.crypto.testing.NullPublicKey
 import net.corda.core.toFuture
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.WireTransaction
