+++ test-utils/src/main/kotlin/net/corda/testing/TransactionDSLInterpreter.kt
@@ -1,6 +1,7 @@
 package net.corda.testing
 
 import net.corda.core.contracts.*
+import net.corda.core.contracts.testing.DummyContract
 import net.corda.core.crypto.SecureHash
 import net.corda.core.identity.Party
 import net.corda.core.seconds
